feat: github action to return the heroku application name for an env
Given an environment, this action will return the Heroku Application Name.

It will look first at the .repo-metadata file to see if an application name
has been defined for the given environment.  If so, it will return that value.
Otherwise, it will return the standard: invh-<application_name>-<environment>

jira: CAPI-96