feat: create unique tag for commits to main
Our plan is to use GitHub workflows to perform push button deployments
via the CI/CD Dashboard.  The workflows will be initiated by creating
a workflow dispatch event via the GitHub api.  These events require
either a branch or a tag as the input to the workflow.  This change updates
the auto-deploy workflow to create a tag for each of the commits to main.

These tags will then be used to trigger deployments and move software
versions to the various enviroments in Heroku

jira: CAPI-151