feat: add 'include-build-metadata' input to the create-git-tag action
The create-git-action action will create a tag in git base upon the version
found in package.json.

Previously, this action would always include a build-metadata suffix in the
generated tag (pre release tags).  This change will update it so that this
action will only include that metadata if the include-build-metadata input is
set to true.

This will allow us to re-use this action for creating production release
tags as well.

jira: ENG-229